Add files via upload
Started on the creation of the basic algorithm. Got through the alpha cost matrix set up which i ended up calling the similarity matrix. Got through most of the Needleman–Wunsch algorithm, semi-stumped on why the wiki has the algorithm optimizing the maximum value(possibly due to the ability to look forward and backward in the sequence or something).